backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Images & Files
Pillow==10.1.0
qrcode[pil]==7.4.2
python-magic==0.4.27

# Excel
pandas==2.1.3
openpyxl==3.1.2

# HTTP Client
httpx==0.25.1
aiohttp==3.9.1

# Validation
email-validator==2.1.0

# Utils
python-slugify==8.0.1
pytz==2023.3

# Logging
python-json-logger==2.0.7
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
